fix: 데이터베이스 스키마 변경
데이터 타입 변경
기온, 최고기온, 최저기온: 정수형 -> 실수형

+++ modules-external/module-mysql/src/main/java/org/example/bigs/pretest/mysql/entity/ForeCast.java
@@ -34,9 +34,9 @@ public class ForeCast {
     @Enumerated(EnumType.STRING)
     @Column(columnDefinition = "VARCHAR(255)")
     private SkyStatus skyStatus;
-    private short temperature;
-    private short maxTemperatureOfDay;
-    private short minTemperatureOfDay;
+    private Float temperature;
+    private Float maxTemperatureOfDay;
+    private Float minTemperatureOfDay;
     private float horizontalWindSpeed;
     private float verticalWindSpeed;
     private float waveHeight;
